Refactor:	Introduce the notion of "plug-in focus" as a means to deal 		with multiple plug-ins desiring input from overlapping 		sources.  The inf::Plugin interface has been made more 		fine grained so that plug-in state is only updated when 		the plug-in has focus while still allowing plug-in actions 		to be performed regardless of focus state.  The idea of 		updating state is primarily geared towards getting the 		state of input devices prior to performing any actions 		based on that information.  Furthermore, the changes to 		inf::Plugin allow all plug-ins to see a consistent 		version of the system state prior to performing their 		operations.  It is certainly my hope that this aspect will 		be a feature rather than a source of bugs.
I am not sure that this is the best way to attack the problem of
overlapping input handling among plug-ins, but it does make it possible
to fix the current issues of combining navigation and grabbing plug-ins
only to see undesirable results.  To take full advantage of this capability,
however, the mode switching plug-in must be used.  The viewer (inf::Viewer)
still executes all its plug-ins and performs no focus management.  The
mode switching plug-in, on the other hand, only performs the state update
on the focused plug-in while still allowing all plug-ins known to it to get
processing time for their specific operations.

Note that there are not really any functional changes being applied to the
current set of navigating and grabbing plug-ins.  Rather, their existing
functionality has just been split up into the new methods for updating
their state and performing their actions.  The two navigating plug-ins do
keep track of their focus and disable navigation when focus is lost.

If this turns out to be a lame idea or to taint the application design,
it can always be backed out.  Revision control rocks.  :)