Refactor:	Get jccl configuration information on demand from the viewer when initializing plugin.
This makes a major change to how plugins are configured.  This change moves the responsibility for configuration totally 
into the plugin classes.  This removes the interface Plugin::config() that the viewer used to manually call on each plugin 
with their associated configuration information.  Instead, now as part of Plugin::init(), each plugin must go out to find 
and use any configuration information they need.

This change reduces the complexity of viewer and removes some order dependence.  It also can allow for more flexible 
configuration of plugins because they are not tied to just a single config element.  They can read from any elements they 
want and in the future it would be possible to add other types of configuration to the configuration object (ex: command 
line map using boost::program_options).

Note: Any plugins that have been written using the old method must be refactored to configure from init.  See the included 
plugins for a method to do this relatively cleanly using the existing code by just having init look up the config element 
to use and then calling the "old" config method with it.