Attempt to make use of a shader written by Terry to highlight the selected object.  Unfortunately, I do not know for sure if this is working as OpenSG is telling me that my PowerBook doesn't support GLSL--probably because VR Juggler is using X11 instead of Cocoa.
This change includes some weak support for falling back on the old
bounding box highlighting method if the shaders cannot be loaded.  It does
not handle the case of GLSL being unavailable because I do not yet know
how to test for that.  The duality of the highlighting mechanisms probably
makes the overall implementation less than acceptable, but I do not have
enough experience with any of this to know how to do it well.