This change (coupled with the changes made to vrj::Kernel yesterday) should fix the crash-on-exit problems.  As far as I can tell, what was going wrong was related to OpenSG objects being deleted in the wrong thread.  This is a little tricky because of all the reference counted memory being used.
A change very similar to this could probably be made to vrj::OpenSGApp
to provide a well-defined mechanism for OpenSG-based application objects
to release their OpenSG memory at the right time.  If vrj::OpenSGApp
were to be modified, then this class would obviously be updated to use
that scheme instead of rolling its own.