Major refactoring of ModeSwitchPlugin
The change adds the ability to have multiple sub-plugins active at the same time in the same mode.
The mode switch plugin is no configured by adding a list of plugins and an associated list of mode ids where the given plugin is supposed to be 
active.

This change is much more a refinement of the current method then it is what I think we will really end up with in the long term.  That said it 
should work for the projects and requirements that we currently have.