Copy over SlaveViewer into the main IOV build and install it along side the standard viewer.
This does not update the vc71 build because I don't have a build environment setup yet on windows.  Since it builds much like the viewer app, I 
don't think this will be an obstacle though.

Once I get approval from the other users that this is a good idea, we can remove the old location of the SlaveViewer.