- Moved SlaveViewer class into IOV library
This will allow custom slave applications to share code with the base class handling of networking and communication.

This change also moved the "exitcodes.h" file contents into IOV/ExitCodes.h to allow SlaveViewer.cpp to return the same exit codes.  It makes a lot of sense to 
have IOV-based applications return common exit codes, so this probably isn't a bad thing.  It may be though that this file should be moved/renamed/etc in the 
future.  I don't have a good way to judge this, so for right now we will just put it here and start using it.