- Added a template method pattern to the viewer and slave viewer to allow for data communication between the viewer and slaves each frame.
The methods allow the master to write/read data from the slaves and allows the slaves to do the same in return.

Unfortunately, the part of the code to handle reading data back from the slaves is not working correctly, so it is commented out until I get 
feedback from the OpenSG mailing list about what to do.  

But for now at least we do have the ability to send data from the master.  I guess that is something.