- Fixed bug where rotation was nearly constant no matter the angle of the wand.
I think this bug had to do with the way a Quat was being set as if it was an AxisAngle.  I just replaced this with an AxisAngle and things seem to 
work fine.

- Added rotation sensitivity property

This property allows user configuration of rotation scalar.  This can help out in cases where the user wants very fine control over rotation.