Extend the wand navigation plug-in to allow users to specify multiple buttons to activate the various actions of the plug-in.  The button numbers are specified using a simple comma-separated list, and the states of the buttons are accumulated using an AND operation.  This means that all the buttons must be in the same state in order for the action to be activated. Hence, this is a pretty brain-dead approach, but it gets things moving in the right direction--more or less.
I have a more advanced implementation of this that allows users to choose
the button state (ON, TOGGLE_OFF, etc.) along with the button numbers, but
I thought that this might get too complicated.  I was also fairly sure
that the plug-in code would not support users setting things up so that,
for example, forward navigation would be toggled on and off instead of
always being on when the button is pressed.