Added:	Initial support for status output 	Initial support for status panel.
This is a pretty major change.  It introduces a new status message output framework for IOV.
The framework is simple stupid right now and will need more work in the future, but
it will work for now and give us what we need.

To output text, we can now use "IOV_STATUS << .....". This will output the text through
the IOV Status singleton which will in turn redirect it to whatever outputters are
registered.

This change adds a StatusPanelPlugin that loads and configures a graphical status
panel.  This panel is registered to get status messages from IOV and it can also be
called directly to set mode and control information on the panel for the active plugins.