- Add scene data type that just holds a shared ptr to the status panel plugin
This smells a little bad to me, but there isn't a good way to do this.  We made the design in such a way that we try to "hide" the plugins from one another 
and there is no central plugin list so we can't look up one plugin from another.  This isn't necessarily bad... but it means that when we want to share a 
plugin we have to do it explicitly through a scene data.  This actually may be a good idea because it keeps all data exchange in a central location (which is 
what we wanted anyway).

- Add code to mode switch plugin to update the panel

Now that we have the panel we have to figure out what to do with it. :)

As a first step I just added some code to the mode switch plugin that will update the header text with the current mode name and info.  This seems useful and 
is thus probably a good place to start.