Extend inf::StatusPanel with what will no doubt be temporary capabilities that allow plug-ins to display their button controls.  I spent two days trying to come with some sort of solution for this very complex problem, and what I have come up with is neither clever nor robust nor scalable. The only good thing about it is that it works.  Since the assembly application has to be delivered in four days, functionality is my primary concern, not good design (as if I could come up with one anyway).  Allen will no doubt come up with a much better solution, and in the meantime, I can continue searching for my Pink Floyd hat.
As part of this change, I modified the signature of the method
inf::Plugin::focusChanged() so that it now takes a parameter of type
inf::ViewerPtr.  All plug-ins that override inf::Plugin::focusChanged()
have to be updated for this change or else their override will not be
called.  The plug-in interface version is now 1.1.

Bumped IOV version to 0.6.0.