For objects that are accessed only through shared pointers and are initialized with an init() method, the init() method now returns the shared pointer to the object being initialized. This allows for the following construction/initialization form:
   inf::MyClassPtr obj = inf::MyClass::create()->init(...);

This is modeled after the conventions used in Objective-C which has no
formal constructor for classes.

The use of boost::shared_ptr<T>, however, presents a complication. For
cases where a virtual init() method is declared in a base class, a covariant
return type cannot be used on a derived type that overrides the base class
init() method. This is because instantiations for boost::shared_ptr<A> and
boost::shared_ptr<B> (where B derives from A) can only be handled
polymorphically through features built into boost::shared_ptr<T>.  In
other words, boost::shared_ptr<B> is not a subclass of boost::shared_ptr<A>,
and as such, the criteria for covariant return types are not met.

Roll version to 0.21.0.