Attempt to work around a corner case of the ray/volume intersection algorithm used by this strategy. When the origin of the ray is inside of or on the edge of the object volume, the enter value returned by OSG::Volume::intersect() is 0.0. Since the algorithm is looking for the closest match (the smallest positive value for min_dist), the first volume that the wand is inside will end up being the match even with triangle intersections enabled.
This change fixes that, but there is another problem. Namely, if the
wand is inside a volume (call it A) that contains no other volumes, a valid
intersection could be ignored. If A is inside the volume of B and B is in
the list of provided objects before A, then the intersection with B will be
the one that gets returned even though the closer match is with A. This
stems from this algorithm being limited by the order of the intersection
tests and needs to be addressed separately.