Instead of having inf::User update its view platform once per frame after the plug-ins may (or may not) have moved the view platform, inf::User now responds to the "platform moved" signal emitted by inf::ViewPlatform to update the platform whenever it moves. This will allow one plug-in to move the view platform when it is updated and then have subsequently updated plug-ins get up-to-date data about the state of the view platform.
This change is part of an effort that I started over a year ago to fix a
bug related to consistent state information between plug-ins. The bug is that
one plug-in may change something about the run-time state but another plug-in
will not see that change until the next frame. An example of this arises
when a navigation plug-in and a grab plug-in are active at the same time. The
navigation plug-in will move the view platform, but since the view platform
was not updated until *after* all plug-ins were updated, the grab plug-in
would always be a frame behind in its positioning of the grabbed object
while navigating.

In Revision 62, I split the method inf::Plugin::update() into
inf::Plugin::updateState() and inf::Plugin::run() and introduced the notion
of plug-in focus. That (much more complicated) change more or less addressed
the specific problem of navigating and grabbing at the same time, though it
never really seemed like the correct approach. This revision may still not
be the correct approach, but it does eliminate the need for having the two
separate methods in inf::Plugin.