Revert the most significant change of Revision 62 by combining inf::Plugin::updateState() and inf::Plugin::run() back into inf::Plugin::update(). The idea of plug-in focus still exists, but now it is up to the override of inf::Plugin::update() to make proper use of the focus state information.
To update the plug-ins for this API change, I simply moved their old
updateState() implementation into an if block that tests the result of
inf::Plugin::isFocused() and then put the run() implementation immediately
after the if block. If I have done this correctly for all the plug-ins,
then there should not be any major behavioral changes when plug-ins are
executed.

Bumped the IOV plug-in API version to 1.2.
Rolled the IOV version to 0.27.0.