Refactored inf::GrabPlugin and inf::GrabStrategy so that it is now the responsibility of the grab stategy plug-ins to keep track of what objects are currently grabbed. This was the case before, though only informally. Now, it is a strict requirement. This approach works better than having inf::GrabPlugin keep track since the variability of object grabbing and releasing is supposed to be encapsulated by the grab strategy.
As part of this change, inf::GrabSignalData has been extended to include a
new signal that can be used for "asynrhonous" object release notification.
inf::SignalGrabStrategy needs to keep track of what objects are grabbed,
and it needs a way to be informed when a grabbed object is released without
the user having invoked the release operation. The new signal in
inf::GrabSignalData allows for that to happen. See the comments for
inf::GrabSignalData::asyncRelease for more information.

Bumped the IOV version to 0.31.0 and the inf::GrabStrategy API version
to 1.1.