The "slide epsilon" concept was pretty badly broken. The information about configuring it stated that the epsilon was measured in feet, but the handling of it was definitely not operating in feet. Instead, the epsilon was being multiplied by the application-specific draw scale factor and then compared against the data as read from the analog input source. Proper physical units were not being factored in at all.
Ultimately, I don't think that the use of physical units for the slide
threshold would work out very well. For example, if the slide threshold were
set to three inches (meaning that the grabbed object would have to move
three inches before it would slide at all), then an object grabbed within
three inches of the wand could never be slid to the point where it was
exactly at the wand position.

So, this change formally defines the slide threshold as being in terms of
the data read from the analog input device. This means that the value must
be in the range (0.0,1.0). All of this is noted in the help information for
the configurable property.