Instead of naming debug-enabled plug-ins with "_d", put them in the 'debug' subdirectory of the directory containing the optimized plug-ins. This makes it so that we can mix a debug IOV-based application with an optimized IOV build. Plug-ins on Windows compiled against the debug runtime will still have the "_d" in their name and be installed next to the optimized plug-ins as this seems to be the widely used convention on Windows.
Note that this change is in line with naming and installation conventions for
VR Juggler 2.2, but this change in IOV does not make VR Juggler 2.2 the
minimum required version. If we were using vpr::LibraryLoder instead of
vpr::LibraryFinder, things would be different.

Bumped the version to 0.35.0. There are no API changes, but this change
will require IOV-based applications to update how they name and install
their plug-ins.