inf::plugin::buildSearchPath() now looks in the 'debug' subdirectory for optimzied and debug-enabled builds. Previously, it only searched the 'debug' subdirectory for debug-enabled builds not compiled against the Visual C++ debug runtime. This change has been made to break the dependency of using a debug-enabled build of IOV with a debug-enabled application based on IOV. Prior to this, it was helpful (read "necessary") to match optimized/optimized and debug/debug to get the correct plug-in search behavior (even after the refactoring that I did in version 0.35.0).
Related to this, inf::PluginRegistry::addEntry() now ignores duplicate
registration attempts. This works around an issue of finding optimized and
debug-enabled versions of the same plug-in (that lack any qualifier to
distingush one from the other). Now, only the first one discovered is added
to the registry.

I have made this change in the interest of keeping IOV's plug-in discovery
as automated as possible. Originally, I had thought (and suggested) that we
leave these decisions up to the configuration. If we did that, the C++
implementation of plug-in searching would be easier. However, it seems like
the software becomes harder to use because the configured plug-in search
path has to be maintained correctly depending on which build of IOV is being
loaded at run time.

Bumped the version to 0.40.0.