Completely revamed inf::DigitalCommand so that it is now something useful rather than a cheesy attempt to be real code. Digital commands are now configured using button state expressions described using a fairly simple language based on boolean logic. inf::DigitalCommand uses a Spirit-based parser to compose functions that test the state of the button(s) identified in the expression. This is something that I have been wanting to do for years, and I am really happy to have this implemented--especially so simply thanks to boost::function<T> and Spirit.
Ultimately, what this means is that code can be written that has no hard
coding of digital proxy names/aliases or digital button states. Plug-ins that
were already using inf::DigitalCommand have been updated for these new
capabilities, and they are noticably simpler than they used to be. There is
still more room for improvement (especially with respect to the use of the
Status Panel... ugh).

Bumped version to 0.41.0.