Renamed the signal objects in inf::StatusPanelData to use the method naming convention. This makes them look just like methods when the signal is emitted. The names, while very similar to what they were before, also evoke more of a sense of an action being associated with the object.
Aside from this, the query signal for determining if the status panel already
contains some control text has had its signature changed so that it returns a
result "out the back" rather than through a by-reference parameter. This makes
usage of that signal much cleaner.

I still can't stand how the Status Panel works, but this is a small
improvement.